requirements.txt
django
djangrestframework
django-cors-headers
djangrestframework-simplejwt
